[Kernel] Phase 4 (communication): move JIT communication operators into kernels.ops.communication (RFC #29630)
Moves all_reduce mp from sglang.jit_kernel into sglang.kernels.ops.communication (classified);
compat shims at old paths (removed Phase 5); csrc stays in jit_kernel.
